+++ app/src/main/java/com/example/dessertclicker/DessertViewModel.kt
@@ -2,6 +2,7 @@ package com.example.dessertclicker
 
 import androidx.lifecycle.ViewModel
 import com.example.dessertclicker.data.Datasource
+import com.example.dessertclicker.model.Dessert
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
@@ -32,5 +33,29 @@ class DessertViewModel: ViewModel() {
                 revenue = revenue
             )
         }
+    }
+
+
+    /**
+     * Determine which dessert to show.
+     */
+    private fun determineDessertToShow(
+        desserts: List<Dessert>,
+        dessertsSold: Int
+    ): Dessert {
+        var dessertToShow = desserts.first()
+        for (dessert in desserts) {
+            if (dessertsSold >= dessert.startProductionAmount) {
+                dessertToShow = dessert
+            } else {
+                // The list of desserts is sorted by startProductionAmount. As you sell more desserts,
+                // you'll start producing more expensive desserts as determined by startProductionAmount
+                // We know to break as soon as we see a dessert who's "startProductionAmount" is greater
+                // than the amount sold.
+                break
+            }
+        }
+
+        return dessertToShow
     }
 }
